requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
mssql-django==1.5
pyodbc==5.2.0
python-dotenv==1.0.1
sqlparse==0.5.3
